+++ modules/globebrowsing/tile/tileprovider/defaulttileprovider.cpp
@@ -96,7 +96,7 @@ DefaultTileProvider::DefaultTileProvider(const ghoul::Dictionary& dictionary)
     int tilePixelSize = 0;
     if (dictionary.getValue<double>(KeyTilePixelSize, pixelSize)) {
         LDEBUG("Default pixel size overridden: " << pixelSize);
-        tilePixelSize = pixelSize; 
+        tilePixelSize = static_cast<int>(pixelSize); 
     }
     TileTextureInitData initData(LayerManager::getTileTextureInitData(
         _layerGroupID, tilePixelSize));
